Progress Report One
I have completed the hero section of the webpage. I believe that I made it well. Each component of the webpage has its own css file and then there will be a master css file that stores all of the css files together. After each commit the css file will be updated since this file will be used as the main file for the project.